+++ src/libkern64/stdio.h
@@ -1,5 +1,5 @@
 #include <shared/screen.h>
 
 extern void print64(const char * buffer);
-extern void printHex64(unsigned int number);
+extern void printHex64(unsigned long number);
